Validate before dynamic render
Such a simple code snippet **could** provided the ability for an attacker to read our
source code and application configuration values.

Or depending the ability of the attacker, it could use this vulnerability to
traverse directory on the server.

So validating the template before it renders it is the secure way to
dynamic render the templates.

Other observation is that on the code you can see that I added
the render body: nil, status: 404. That's because we don't have at
the moment a 404 page.