Add hero image and hero description component
* Add hero image component
* Add hero description component
* Fix a bug on cta component when the page does not have a cta link.
* Update the cassettes in order to get the nodes of hero image and hero
* Fix a bug on feedback email. When the email has <p> tags the mail to opens
  with the <p> tags in the from field, so strip tags solves the issue.
  description

The image is a background image because of the way we require it to
scale across devices. An image tag would result in a image which is
always confined within the container it's in, not ideal for mobile view
where the image would simply shrink.

Using a background image means that we can simply adjust the
background position and scale it to fit either the height or width of
the container element (but not both). This provides us with extra
flexibility on which images we use, and fixes the issue of varying
screen widths.