Make feature step to be more consistent
Before if the step fails the following were returned:

expected "evidence-types__item-icon
evidence-types__item-icon--tick svg-icon svg-icon--evidence-tick" to include
"evidence-types__item-icon--cross"

This message is not so intuitive for any developer that will face this.

So then this solution grep the "tick" or the "cross" from the html class
and have a simple equality expectation. The result:
   expected: "cross"
   got: "tick"