Add evidence type and evidence checlist to Insight & Evaluation
So now the evidence hub controller inherits from evidence summaries controller
altogether with Insight and Evaluation.

Also refactor the evidence hub index view to have a search form partial as
well as search result.

So now to add the evidence checklist you only need to add the values on the
en.yml locale and by the page type and it is because it is using template
inheritance from evidence summaries everything will work out of the box.