Add pagination to the evidence hub
Move the form objects to app/forms to become more consistent
the places to look for classes.

Moved the evidence hub search form params method to be public
and acessible by the presenter to generate the link to go to the
next page and previous page.

Re-recording all cassettes to send the per page 20 as default as well.