Update mas_cms_client setup var to use FRONTEND_REQUEST_TIMEOUT
This could well be a temporary fix. Mainly to try to get
fincap preview up and working as it currently is missing the
timeout variable. By using the same as Frontend it means not having
to add another env variable to the build.

Improvements will come on the mas-cms-client. Where a default timeout
should be set. Errors should also be improved as this was not directly
obvious that the missing variable was the reason of the error