Add thematic review template
The thematic review templates is exactly like an article so
I create a structure which both Article and Thematic review uses.

I created a folder called app/templates. Basically a template
is a page which contain "n" components.

I also using the Rails template inheritance feature which means
both Article and Thematic Review controller are using the same
template, e.g:

    class ArticlesController < FincapTemplatesController
    end

    class ThematicReviewsController < FincapTemplatesController
    end

    app/views/
    ├── fincap_templates
    │   └── show.html.erb

The show templates uses the presenter for the resource object.
The presenter for all templates I decided to name TemplatePresenter
which will be responsible for rendering each component from the project.

So in order to add more other page types that looks like an article:

1) Add the feature test
2) Create the controller inheriting from FincapTemplatesController
3) Create the presenter inheriting from TemplatePresenter