Experiment with mapping to Document:
It would make sense for ThematicReviewsLandingPage to inherit
from the Article resource as it is similar to ThematicReview. However,
the `all` method for the Article resource in mas-cms-client is
not working. NOTE: The `find` method for Article does work.

This approach maps ThematicReviewsLandingPage and ThematicReview
to Document and uses a DocumentPresenter for displaying the results.