Sort news by published date
The "Latest news" component doesn't show the
most recent news items as defined by the "order by date" field. Instead, it
currently shows the first 3 created news items in CMS.

The way that I solved was to add the sort feature to TaggedNews class.

The TaggedNews class is being used in all pages that the bug appears:

  Homepage
  Lifestages
  Thematic reviews
  Country pages

Since the class is not manipulating data I kinda need to make a
published date method which I strip the tags and the carriage
returns and parse as Date. In order to not break compatibility
with the existing behaviour of all those pages that this class is being used
I needed to make this class inherits from simple delegator.