Add filter by Evidence type on Evidence Hub
We want to be able to filter the Evidence Hub results by the evidence
type.
This commit is assuming that in case of not selecting any of them,
the user will get the results including all the types.

In the Evicence Hub Controller we have two groups of parameters:
- The params used for the formulary filters.
- The params used for searching the CMS.

In case of the Evidence Type, we want it as parameter for the form
filters, but it cannot be used for the search in CMS without being
mapped to the 'document_type' parameter in the search to CMS.

This is caused by Evidence Types being a subset of CMS Document Types,
while they don't directly exist as a resource in CMS.

If we send the evidence types as a search parameters for filtering the
search results, it will return no documents at all as won't be able
to find any results by 'evidence_type' key.