Wrap & extract year of publication parsing logic
Reading the Search For Param Parser I felt that there were multiple
constants and methods exclusively related to parse the years of
publication filter.

As the Search Parser has some data structures transformations that
make the reader needing to spend some time through the code to figure
out what is happening (or at least I needed it), I decided to try
to leave this class responsibility as focused as possible in these
data structures transformations, trying to achieve an easier to read
and mantain class.

In order to achieve that I grouped and extracted the Year Of
Publication parsing logic into a module used by the Search Form Param
Parser.