Set a page title for specific pages
The site uses the gem meta-tags which allows you to set
a specific title on the page which is used in the title tag.

For the majority of pages requiring a specific page title, we
can use the translations to set the title.

The thematic_reviews index page requires a different approach for
defining the page title because it makes use of the template views.
Therefore, I have created a method in the template presenter to define
the page title.